Fix hassio circular reference (#6204)
